requirements.txt
streamlit
torch
torchvision
Pillow
pyttsx3
gtts
gdown